GDB-11856 fix compilation of dynamic form field directive in production build
