+++ src/js/angular/core/directives/dynamic-form/field-type.js
@@ -1,0 +1,13 @@
+/**
+ * Enum for field types.
+ * @readonly
+ * @enum {string}
+ */
+export const FIELD_TYPE = {
+    STRING: 'string',
+    TEXT: 'text',
+    BOOLEAN: 'boolean',
+    JSON: 'json',
+    SELECT: 'select',
+    MULTI_SELECT: 'multi_select'
+}
